pyproject.toml
[project]
name = "torchtmpl"
version = "0.0.1"

dependencies = [
    "PyYAML>=6.0",
    "torchinfo>=1.7.2",
    "wandb>=0.13.11",
    "matplotlib>=3.7.1",
    "torchcvnn>=0.3.1",
    "scikit-image>=0.22.0",
    "torchvision>=0.16",
]

[options]
package_dir = "torchtmpl"
